Prepare #294 Pull up DefaultDependenciesTool into DependenciesTool
